requirements.txt
streamlit
torch
transformers
pdfplumber
python-docx